+++ changelog.md
@@ -301,3 +301,5 @@
 * Added Adventurers
 * Villager now chooses the best equipment
 * Added more eye variants
+* Fixed zombies not always using zombie clothing
+* Villager on fire will now burn their clothes